sentiment-analysis/requirements.txt
sense-ai-sdk
chaice